More cleanup after removing legacy protocol code
