+++ src/extension/analysis/analyzer_lsp.ts
@@ -1,3 +1,4 @@
+import * as fs from "fs";
 import * as path from "path";
 import * as stream from "stream";
 import * as vs from "vscode";
@@ -14,7 +15,7 @@ import { DartToolingDaemon } from "../../shared/services/tooling_daemon";
 import { fsPath } from "../../shared/utils/fs";
 import { ANALYSIS_FILTERS } from "../../shared/vscode/constants";
 import { DartTextDocumentContentProviderFeature } from "../../shared/vscode/dart_text_document_content_provider";
-import { cleanDartdoc, createMarkdownString } from "../../shared/vscode/extension_utils";
+import { cleanDartdoc, createMarkdownString, extensionVersion } from "../../shared/vscode/extension_utils";
 import { InteractiveRefactors } from "../../shared/vscode/interactive_refactors";
 import { CommonCapabilitiesFeature } from "../../shared/vscode/lsp_common_capabilities";
 import { LspUriConverters } from "../../shared/vscode/lsp_uri_converters";
@@ -26,8 +27,7 @@ import { checkForLargeNumberOfTodos } from "../user_prompts";
 import { reportAnalyzerTerminatedWithError } from "../utils/misc";
 import { safeToolSpawn } from "../utils/processes";
 import { getDiagnosticErrorCode } from "../utils/vscode/diagnostics";
-import { getAnalyzerArgs } from "./analyzer";
-import { SnippetTextEditFeature } from "./analyzer_lsp_snippet_text_edits";
+import { SnippetTextEditFeature } from "./analyzer_snippet_text_edits";
 import { FileTracker } from "./file_tracker";
 
 export class LspAnalyzer extends Analyzer {
@@ -481,7 +481,7 @@ export class LspAnalyzer extends Analyzer {
 			"dartAnalysisLSP",
 			"Dart Analysis Server",
 			async () => {
-				const streamInfo = await this.spawnServer(logger, sdks, dartCapabilities);
+				const streamInfo = await this.spawnServer(logger, sdks);
 				const jsonEncoder = ls.RAL().applicationJson.encoder;
 
 				return {
@@ -574,11 +574,11 @@ export class LspAnalyzer extends Analyzer {
 		return client;
 	}
 
-	private spawnServer(logger: Logger, sdks: DartSdks, dartCapabilities: DartCapabilities): Promise<StreamInfo> {
+	private spawnServer(logger: Logger, sdks: DartSdks): Promise<StreamInfo> {
 		// TODO: Replace with constructing an Analyzer that passes LSP flag (but still reads config
 		// from paths etc) and provide it's process.
 		const vmPath = path.join(sdks.dart, dartVMPath);
-		const args = getAnalyzerArgs(logger, sdks, dartCapabilities);
+		const args = getAnalyzerArgs(logger);
 
 		logger.info(`Spawning ${vmPath} with args ${JSON.stringify(args)}`);
 		const process = safeToolSpawn(undefined, vmPath, args);
@@ -610,4 +610,63 @@ class LoggingTransform extends stream.Transform {
 		this.push(chunk, encoding);
 		callback();
 	}
+}
+
+export function getAnalyzerArgs(logger: Logger) {
+	const analyzerPath = config.analyzerPath || "language-server";
+
+	// If the ssh host is set, then we are running the analyzer on a remote machine, that same analyzer
+	// might not exist on the local machine.
+	if (!config.analyzerSshHost && analyzerPath !== "language-server" && !fs.existsSync(analyzerPath)) {
+		const msg = "Could not find a Dart Analysis Server at " + analyzerPath;
+		void vs.window.showErrorMessage(msg);
+		logger.error(msg);
+		throw new Error(msg);
+	}
+
+	let analyzerArgs = [];
+
+	// Optionally start the VM service for the analyzer.
+	const vmServicePort = config.analyzerVmServicePort;
+	if (vmServicePort) {
+		analyzerArgs.push(`--enable-vm-service=${vmServicePort}`);
+		// When using LSP, printing the VM Service URI will break the protocol and
+		// stop the client from working, so it needs to be hidden.
+		analyzerArgs.push(`-DSILENT_OBSERVATORY=true`);
+		analyzerArgs.push(`-DSILENT_VM_SERVICE=true`);
+		analyzerArgs.push(`--disable-service-auth-codes`);
+		analyzerArgs.push(`--no-dds`);
+		analyzerArgs.push("--no-serve-devtools");
+	}
+
+	// Allow arbitrary VM args to be passed to the analysis server.
+	if (config.analyzerVmAdditionalArgs)
+		analyzerArgs = analyzerArgs.concat(config.analyzerVmAdditionalArgs);
+
+	analyzerArgs.push(analyzerPath);
+
+	if (analyzerPath === "language-server") {
+		analyzerArgs.push("--protocol=lsp");
+	} else {
+		analyzerArgs.push("--lsp");
+	}
+
+	// Optionally start the analyzer's diagnostic web server on the given port.
+	if (config.analyzerDiagnosticsPort)
+		analyzerArgs.push(`--port=${config.analyzerDiagnosticsPort}`);
+
+	// Add info about the extension that will be collected for crash reports etc.
+	const clientID = isRunningLocally ? "VS-Code" : "VS-Code-Remote";
+	analyzerArgs.push(`--client-id=${clientID}`);
+	analyzerArgs.push(`--client-version=${extensionVersion}`);
+
+	// The analysis server supports a verbose instrumentation log file.
+	if (config.analyzerInstrumentationLogFile)
+		analyzerArgs.push(`--instrumentation-log-file=${config.analyzerInstrumentationLogFile}`);
+
+	// Allow arbitrary args to be passed to the analysis server.
+	if (config.analyzerAdditionalArgs)
+		analyzerArgs = analyzerArgs.concat(config.analyzerAdditionalArgs);
+
+	return analyzerArgs;
 }
