+++ src/extension/analysis/analyzer_snippet_text_edits.ts
